% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{This is from oxygen skeleton from code}
\usage{
say(what)
}
\arguments{
\item{what}{character string to be printed, if NULL it returns the default one. Help files are generated by this....}
}
\value{
Whaled shaped message
}
\description{
This is from oxygen skeleton from code
}
\examples{
say("Hi falks")
}
